OCaml: Improve matching `*)` insertion
Currently when we type in `(*` in an OCaml file, we automatically insert
a matching `*)` to ease the chore of closing a comment. However, we can
do better. Usually you want to comment a line of code. In that case it
better to go to the end of the line and then add `*)`.

Before:
  (* *) let a = 5
     ^^
     Inserted automatically

Now:
  (* let a = 5 *)
               ^^
               Inserted automatically